feat(rust): move authority node code level above in `ockam_api`
