+++ implementations/rust/ockam/ockam_command/src/authority/create.rs
@@ -7,11 +7,11 @@ use miette::Context as _;
 use miette::{miette, IntoDiagnostic};
 use ockam::identity::{AttributesEntry, Identifier};
 use ockam::Context;
+use ockam_api::authority_node;
+use ockam_api::authority_node::{OktaConfiguration, TrustedIdentity};
 use ockam_api::bootstrapped_identities_store::PreTrustedIdentities;
 use ockam_api::cli_state::init_node_state;
 use ockam_api::cli_state::traits::{StateDirTrait, StateItemTrait};
-use ockam_api::nodes::authority_node;
-use ockam_api::nodes::authority_node::{OktaConfiguration, TrustedIdentity};
 use ockam_api::nodes::models::transport::{CreateTransportJson, TransportMode, TransportType};
 use ockam_api::DefaultAddress;
 use ockam_core::compat::collections::HashMap;
